add alias to work with plugin and update function
Today we did not have any plugins to work with the plugin and it was necessary to write a lot. I created some shortcuts to make it easier and more productive to work with the plugin. I also took the opportunity to create a feature that allows you to update the plugin to find new features